+++ extension/devtools/app/lib/main.dart
@@ -28,6 +28,7 @@ part 'status_styles.dart';
 part 'color_maps.dart';
 part 'effects_timeline_constants.dart';
 part 'palette.dart';
+part 'theme.dart';
 part 'widgets/adaptive_wrap.dart';
 part 'widgets/action_buttons.dart';
 part 'widgets/action_pill.dart';
@@ -127,85 +128,5 @@ class OrefDevToolsScope extends InheritedNotifier<OrefDevToolsController> {
         .dependOnInheritedWidgetOfExactType<OrefDevToolsScope>();
     assert(scope != null, 'OrefDevToolsScope not found in widget tree.');
     return scope!.notifier!;
-  }
-}
-
-class OrefTheme {
-  static ThemeData light() => _build(Brightness.light);
-  static ThemeData dark() => _build(Brightness.dark);
-
-  static ThemeData _build(Brightness brightness) {
-    final baseScheme = ColorScheme.fromSeed(
-      seedColor: OrefPalette.teal,
-      brightness: brightness,
-    );
-    final scheme = baseScheme.copyWith(
-      primary: OrefPalette.teal,
-      onPrimary: brightness == Brightness.dark ? Colors.black : Colors.white,
-      secondary: OrefPalette.indigo,
-      onSecondary: Colors.white,
-      error: const Color(0xFFFF6B6B),
-      onError: Colors.white,
-      surface: brightness == Brightness.dark
-          ? const Color(0xFF141B22)
-          : const Color(0xFFFFFFFF),
-      onSurface: brightness == Brightness.dark
-          ? const Color(0xFFEAF2F8)
-          : const Color(0xFF11161D),
-    );
-
-    final baseTextTheme = brightness == Brightness.dark
-        ? ThemeData.dark().textTheme
-        : ThemeData.light().textTheme;
-    final textTheme = GoogleFonts.spaceGroteskTextTheme(baseTextTheme).copyWith(
-      headlineSmall: baseTextTheme.headlineSmall?.copyWith(
-        fontWeight: FontWeight.w600,
-        letterSpacing: -0.2,
-      ),
-      titleLarge: baseTextTheme.titleLarge?.copyWith(
-        fontWeight: FontWeight.w600,
-        letterSpacing: -0.2,
-      ),
-      titleMedium: baseTextTheme.titleMedium?.copyWith(
-        fontWeight: FontWeight.w600,
-        letterSpacing: -0.1,
-      ),
-      titleSmall: baseTextTheme.titleSmall?.copyWith(
-        fontWeight: FontWeight.w600,
-        letterSpacing: 0.0,
-      ),
-      bodyLarge: baseTextTheme.bodyLarge?.copyWith(height: 1.4),
-      bodyMedium: baseTextTheme.bodyMedium?.copyWith(height: 1.4),
-      bodySmall: baseTextTheme.bodySmall?.copyWith(height: 1.3),
-      labelLarge: baseTextTheme.labelLarge?.copyWith(
-        fontWeight: FontWeight.w600,
-        letterSpacing: 0.2,
-      ),
-      labelMedium: baseTextTheme.labelMedium?.copyWith(
-        fontWeight: FontWeight.w600,
-        letterSpacing: 0.2,
-      ),
-      labelSmall: baseTextTheme.labelSmall?.copyWith(
-        fontWeight: FontWeight.w600,
-        letterSpacing: 0.4,
-      ),
-    );
-
-    return ThemeData(
-      brightness: brightness,
-      colorScheme: scheme,
-      textTheme: textTheme,
-      scaffoldBackgroundColor: Colors.transparent,
-      useMaterial3: true,
-      dividerColor: brightness == Brightness.dark
-          ? const Color(0xFF24313B)
-          : const Color(0xFFE1E6EC),
-      appBarTheme: const AppBarTheme(
-        backgroundColor: Colors.transparent,
-        elevation: 0,
-        scrolledUnderElevation: 0,
-      ),
-      cardTheme: const CardThemeData(color: Colors.transparent, elevation: 0),
-    );
   }
 }
